switch from Long to Short

+++ spark/F.scala
@@ -3,7 +3,7 @@ package pure
 object Main {
   def main(args: Array[String]): Unit = {
     val modulo = 3
-    val nodeSeq = Seq(0,1,2,0).map(_.toLong)
+    val nodeSeq = Seq(0,1,2,0).map(_.toShort)
     val chordSeq = F.toChordSeq(modulo)(nodeSeq)
     assert(chordSeq == Seq(1,1,1))
     assert(F.lessThanSeq(Seq(1,2,3), Seq(2,3,1)))
@@ -24,37 +24,37 @@ object Main {
 }
 
 object F {
-  def otherNodes(modulo: Long): IndexedSeq[Long] = otherNodes(modulo.toInt)
-    .map(_.toLong)
+  def otherNodes(modulo: Short): IndexedSeq[Short] = otherNodes(modulo.toInt)
+    .map(_.toShort)
 
   def otherNodes(modulo: Int) = Range(1, modulo)
 
-  case class Partition(i: Long, of: Long) {
-    def values(modulo: Long) = otherNodes(modulo)
+  case class Partition(i: Short, of: Short) {
+    def values(modulo: Short) = otherNodes(modulo)
       .filter(_ % of == i)
   }
 
-  def possiblePaths(modulo: Long)(values: Iterator[Long]): Iterator[Seq[Long]] = values
+  def possiblePaths(modulo: Short)(values: Iterator[Short]): Iterator[Seq[Short]] = values
     .flatMap(possiblePaths(modulo, _))
 
-  def possiblePaths(modulo: Long, first: Long): Iterator[Seq[Long]] = otherNodes(modulo)
+  def possiblePaths(modulo: Short, first: Short): Iterator[Seq[Short]] = otherNodes(modulo)
     .filterNot(_ == first)
     .permutations
     // rotating one position avoids partition skew later in isCanonical
     .map(first +: _ :+ 0L :+ first)
 
-  def toChordSeq(modulo: Long)(wrappedCycle: Seq[Long]) = wrappedCycle
+  def toChordSeq(modulo: Short)(wrappedCycle: Seq[Short]) = wrappedCycle
     .sliding(2)
     .map(x => (modulo + x.last - x.head) % modulo)
     .toSeq
 
   // assumes same length
-  def lessThanSeq(a: Seq[Long], b: Seq[Long]) = a.lazyZip(b)
+  def lessThanSeq(a: Seq[Short], b: Seq[Short]) = a.lazyZip(b)
     .dropWhile(x => x._1 == x._2)
     .headOption
-    .exists{ case(c: Long, d: Long) => c < d}
+    .exists{ case(c: Short, d: Short) => c < d}
 
-  def isCanonical(modulo: Int)(chordSeq: Seq[Long]) = !Iterator
+  def isCanonical(modulo: Int)(chordSeq: Seq[Short]) = !Iterator
     .continually(chordSeq)
     .flatten
     .drop(1) //don't start with the current chordSeq!
@@ -63,7 +63,7 @@ object F {
     //.take(modulo) //safety against infite loops (depending what's fed in)
     .exists(lessThanSeq(_, chordSeq))
 
-  def possiblyCanonical(chordPartSeq: Seq[Long]) =
+  def possiblyCanonical(chordPartSeq: Seq[Short]) =
     !chordPartSeq
       .tails
       .drop(1) // don't start with the current chordPartSeq!
@@ -72,9 +72,9 @@ object F {
 }
 
 /*
-object SeqOrdering extends Ordering[Seq[Long]] {
+object SeqOrdering extends Ordering[Seq[Short]] {
   // assumes same length
-  def compare(a: Seq[Long], b: Seq[Long]) = {
+  def compare(a: Seq[Short], b: Seq[Short]) = {
     val toCompare = a.zip(b)
       .toIterator
       .dropWhile{ case (a, b) => a == b }
